chore: configure prettier
So prettier was installed, but I guess you did not have a plugin yet for
your editor - I can definitely recommend installing one. I've added a
configuration file for it as well with some preferences that I
personally have.

There are a quite a few ways in which you can configure it, to see the
options: https://prettier.io/docs/en/options

A neat thing about prettier, is that it also respects the settings in
your .editorconfig. Ideally, you have a single source of truth for your
code formatting configuration, so since we already defined indentation
style and such in the .editorconfig, we don't also want to define it in
the prettier config.