chore: configure eslint
Alright so eslint was already in place, I converted the config to json,
cause personal preference, definitely not a requirement.

Besides that, I added both the config and the plugin for
eslint-prettier. This makes sure that any code formatting that is not
according to prettier's settings will show up as errors through eslint.